Create 251015_SA_vol_3cm_exploration.R
This R script contains the code to run the in silico model presented in: Supporting materials III – Mathematical model exploring role of aerobic glycolysis in 3cm-rule.
This code also generates the figures presented in Suppl. Fig. 7, which visualise the model output.
Note that in this code, we refer to the Immune Dysfunction Score (IDS) as the Metastatic Potential Index (MPI); this metric was renamed in the manuscript but is quantified identically.